added a function to transform a point

+++ src/main/java/com/palyrobotics/util/Point.java
@@ -2,9 +2,9 @@ package com.palyrobotics.util;
 
 public class Point {
 
-    int x, y;
+    double x, y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
